+++ addon/-private/constraints/big-decimal.js
@@ -1,4 +1,5 @@
 import { isPresent } from '@ember/utils';
+import { t } from '@zestia/ember-validation';
 
 const fullPattern = /^-?(\d+)(?:\.(\d+))?$/;
 const decimalPattern = /^(\d*?)0*$/;
@@ -16,22 +17,23 @@ export default function bigDecimal(_options = {}) {
     }
 
     const result = fullPattern.exec(value);
+    const data = { value, ...options };
 
     if (!result) {
-      return 'Value must be a number';
+      return t('big-decimal.must-be-a-number', data);
     }
 
     const integerPart = result[1];
 
     if (integerPart.length > options.maxIntegerDigits) {
-      return 'Value too large';
+      return t('big-decimal.value-too-large', data);
     }
 
     if (result[2]) {
       const decimalPart = decimalPattern.exec(result[2])[1];
 
       if (decimalPart.length > options.maxDecimalDigits) {
-        return `Value must have a maximum of ${options.maxDecimalDigits} decimal places`;
+        return t('big-decimal.max-decimal-places', data);
       }
     }
   };