+++ addon/-private/constraints/date.js
@@ -1,6 +1,7 @@
 import { isPresent } from '@ember/utils';
 import isValid from 'date-fns/isValid';
 import parse from 'date-fns/parse';
+import { t } from '@zestia/ember-validation';
 
 export default function date(options) {
   return function (value) {
@@ -19,6 +20,6 @@ export default function date(options) {
       return;
     }
 
-    return options.message || `Invalid date, expecting ${format}`;
+    return t('date', { value, format });
   };
 }