+++ addon/-private/constraints/present.js
@@ -1,11 +1,12 @@
 import { isPresent } from '@ember/utils';
+import { t } from '@zestia/ember-validation';
 
-export default function present(options = {}) {
+export default function present() {
   return function (value) {
     if (isPresent(value)) {
       return;
     }
 
-    return options.message || 'Required value';
+    return t('present', { value });
   };
 }