+++ tests/unit/constraints/number-test.js
@@ -24,14 +24,6 @@ module('number', function () {
     );
   });
 
-  test('it returns custom message if invalid', function (assert) {
-    assert.expect(1);
-
-    const func = number({ message: 'bad number' });
-
-    assert.strictEqual(func('foo@bar'), 'bad number');
-  });
-
   test('inputs', function (assert) {
     assert.expect(7);