+++ tests/unit/constraints/truthy-test.js
@@ -13,13 +13,4 @@ module('truthy', function () {
 
     assert.strictEqual(truthy()(false), 'Must be truthy');
   });
-
-  test('it returns a custom message if invalid', function (assert) {
-    assert.expect(1);
-
-    assert.strictEqual(
-      truthy({ message: 'please accept the terms' })(),
-      'please accept the terms'
-    );
-  });
 });