+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Change import path for `collateErrors` and `flattenErrors` from
   `@zestia/ember-validation/utils`, to just `@zestia/ember-validation`
+- Removes `message` option from each constraint
+- Makes internationalisation possible by exposing `setMessageFn`, and `t`.
 
 ## 4.4.2